+++ src/Database/migrations/2021_11_29_032839_add_authit_columns_to_users_table.php
@@ -10,7 +10,7 @@ class AddAuthitColumnsToUsersTable extends Migration
     {
         if (!Schema::hasColumn('users', 'avatar')) {
             Schema::table('users', function (Blueprint $table) {
-                $table->string('avatar')->nullable();
+                $table->string('avatar', 2048)->nullable();
             });
         }
     }