+++ src/routes.php
@@ -1,18 +1,16 @@
 <?php
 
-use Naykel\Authit\Http\Controllers\UserController;
-// use Naykel\Authit\Http\Livewire\Profile;
-// use Naykel\Authit\Http\Livewire\UpdatePasswordForm;
-
-
-use Naykel\Authit\Http\Controllers\Auth\EmailVerificationNotificationController;
-use Naykel\Authit\Http\Controllers\Auth\EmailVerificationPromptController;
 use Naykel\Authit\Http\Controllers\Auth\AuthenticatedSessionController;
 use Naykel\Authit\Http\Controllers\Auth\ConfirmablePasswordController;
+use Naykel\Authit\Http\Controllers\Auth\EmailVerificationNotificationController;
+use Naykel\Authit\Http\Controllers\Auth\EmailVerificationPromptController;
+use Naykel\Authit\Http\Controllers\Auth\NewPasswordController;
 use Naykel\Authit\Http\Controllers\Auth\PasswordResetLinkController;
 use Naykel\Authit\Http\Controllers\Auth\RegisterUserController;
-use Naykel\Authit\Http\Controllers\Auth\NewPasswordController;
 use Naykel\Authit\Http\Controllers\Auth\VerifyEmailController;
+use Naykel\Authit\Http\Controllers\UserController;
+use Naykel\Authit\Http\Livewire\Profile;
+use Naykel\Authit\Http\Livewire\UpdatePasswordForm;
 
 use Spatie\Honeypot\ProtectAgainstSpam;
 
@@ -37,8 +35,8 @@ Route::middleware(['web', 'auth', 'verified'])->prefix('user')->name('user')->gr
 
     Route::get('/dashboard', [UserController::class, 'dashboard'])->name('.dashboard');
     // Route::get('/edit-account', [UserController::class, 'edit'])->name('.edit');
-    // Route::get('/edit-profile', Profile::class)->name('.edit-password'); // livewire component
-    // Route::get('/update-password', UpdatePasswordForm::class)->name('.update-password'); // livewire component
+    Route::get('/edit-profile', Profile::class)->name('.edit-password'); // livewire component
+    Route::get('/update-password', UpdatePasswordForm::class)->name('.update-password'); // livewire component
 
 });
 
@@ -58,52 +56,3 @@ Route::middleware(['web', 'auth'])->group(function () {
 
     Route::post('logout', [AuthenticatedSessionController::class, 'destroy'])->name('logout');
 });
-
-
-// FORTIFY
-
-// Route::group(['middleware' => config('fortify.middleware', ['web'])], function () {
-//     $enableViews = config('fortify.views', true);
-
-//     $limiter = config('fortify.limiters.login');
-//     $twoFactorLimiter = config('fortify.limiters.two-factor');
-//     $verificationLimiter = config('fortify.limiters.verification', '6,1');
-
-//     // Two Factor Authentication...
-//     if (Features::enabled(Features::twoFactorAuthentication())) {
-//         if ($enableViews) {
-//             Route::get('/two-factor-challenge', [TwoFactorAuthenticatedSessionController::class, 'create'])
-//                 ->middleware(['guest:' . config('fortify.guard')])
-//                 ->name('two-factor.login');
-//         }
-
-//         Route::post('/two-factor-challenge', [TwoFactorAuthenticatedSessionController::class, 'store'])
-//             ->middleware(array_filter([
-//                 'guest:' . config('fortify.guard'),
-//                 $twoFactorLimiter ? 'throttle:' . $twoFactorLimiter : null,
-//             ]));
-
-//         $twoFactorMiddleware = Features::optionEnabled(Features::twoFactorAuthentication(), 'confirmPassword')
-//             ? [config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard'), 'password.confirm']
-//             : [config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')];
-
-//         Route::post('/user/two-factor-authentication', [TwoFactorAuthenticationController::class, 'store'])
-//             ->middleware($twoFactorMiddleware)
-//             ->name('two-factor.enable');
-
-//         Route::delete('/user/two-factor-authentication', [TwoFactorAuthenticationController::class, 'destroy'])
-//             ->middleware($twoFactorMiddleware)
-//             ->name('two-factor.disable');
-
-//         Route::get('/user/two-factor-qr-code', [TwoFactorQrCodeController::class, 'show'])
-//             ->middleware($twoFactorMiddleware)
-//             ->name('two-factor.qr-code');
-
-//         Route::get('/user/two-factor-recovery-codes', [RecoveryCodeController::class, 'index'])
-//             ->middleware($twoFactorMiddleware)
-//             ->name('two-factor.recovery-codes');
-
-//         Route::post('/user/two-factor-recovery-codes', [RecoveryCodeController::class, 'store'])
-//             ->middleware($twoFactorMiddleware);
-//     }
-// });