+++ changelog.md
@@ -1,15 +1,18 @@
 # Changelog
 
 ## Whats Next?
-- user profile
+- user profile profile picture
 - two factor authentication
 - notifications on login
+- add disk on install command
+- create copy command for local usage
 
 
 ## 1.4.0 (pending)
 
 - remove dependency on fortify
 - add new auth controllers (based on breeze)
+- Add UpdatePassword livewire component and rules
 
 ## 1.3.0